feat(updateUserSettings) - render logic refactored. minor enhacements
