+++ package.json
@@ -29,6 +29,7 @@
     "express": "^4.14.1",
     "express-async-handler": "^1.0.4",
     "express-session": "^1.15.1",
+    "filter-obj": "^1.1.0",
     "history": "^4.6.3",
     "morgan": "^1.8.1",
     "passport": "^0.3.2",
